requirements.txt
astroid==2.4.2
autopep8==1.5.4
colorama==0.4.4
imutils==0.5.3
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.3
opencv-python==4.5.1.48
Pillow==8.0.1
pycodestyle==2.6.0
pylint==2.6.0
pytesseract==0.3.7
python-dotenv==0.15.0
six==1.15.0
tesseract==0.1.3
toml==0.10.2
wrapt==1.12.1
